Extract coop TUI markdown renderer
Committed-By-Agent: codex
